🎯 **15/15 tests PASS. Comprehensive fix verified.** Now let me package this up for you.
# 📋 Audit Summary + Fix Deployment

## What I Reviewed
Systematically checked every `int` field in `backend/models/schemas.py` (~50 fields) for the same class of bug where legacy production data could contain decimal values that crash Pydantic v2 validation.

## What I Found

| Model | Field | Original | Fixed | Why It's Fractional in Real Life |
|---|---|---|---|---|
| `AllocationResponse` | `percentage` | `int` | `float` | ⚠️ 2.5 in your prod (THE bug) |
| `AllocationResponse` | `hours` | `Optional[int]` | `Optional[float]` | Hours like 4.5, 8.75 are natural |
| `AllocationResponse` | `actual_percentage` | `Optional[int]` | `Optional[float]` | Same category as `percentage` |
| `PhaseAllocation` | `percentage` / `hours` | `Optional[int]` | `Optional[float]` | Nested in allocation responses |
| `AllocationCreate/Update` | 3 fields | `int` | `float` | Frontend can input decimals |
| `AllocationValidateRequest` | `percentage` / `hours` | `Optional[int]` | `Optional[float]` | Validator would crash on decimals |
| `MoveResourceRequest` | `new_percentage` | `Optional[int]` | `Optional[float]` | Same |
| `ProjectCreate/Update/Response` | `actual_progress` | `Optional[int]` | `Optional[float]` | Progress can be 33.3% |
| `StatusUpdateCreate/Response/Edit` | `actual_progress` | `int` or `Optional[int]` | `float` or `Optional[float]` | Same |
| `ResourceCreate/Update/Response` | `standard_capacity` | `int` | `float` | Part-time can be 87.5% |

## What I Left as `int` (correctly)
- `WBSTaskResponse.order` — integer sort key
- `ReportLinkResponse.view_count` — counter
- `RescheduleProjectRequest.weeks_to_shift` — count of weeks

# 🚀 Deploy the Fix

## In Your Cloud Shell Repo

```bash
cd path/to/your/dd-planner-repo

# Open the file
cloudshell edit backend/models/schemas.py
```

Because there are 12 lines to change across the file, use this Python script that applies all fixes in one go:

```bash
python3 <<'PYEOF'
path = "backend/models/schemas.py"
with open(path) as f:
    content = f.read()

replacements = [
    # ResourceCreate/Update/Response
    ("standard_capacity: int = 100", "standard_capacity: float = 100"),
    ("standard_capacity: Optional[int] = None", "standard_capacity: Optional[float] = None"),
    ("    standard_capacity: int\n", "    standard_capacity: float\n"),

    # ProjectResponse / ProjectCreate / ProjectUpdate: actual_progress
    ("    actual_progress: Optional[int] = None\n", "    actual_progress: Optional[float] = None\n"),

    # StatusUpdateResponse.actual_progress (required int → required float)
    ("    actual_progress: int\n", "    actual_progress: float\n"),

    # PhaseAllocation.percentage / hours
    ("    percentage: Optional[int] = None\n    hours: Optional[int] = None\n",
     "    percentage: Optional[float] = None\n    hours: Optional[float] = None\n"),

    # AllocationCreate/Update.actual_percentage
    ("    actual_percentage: Optional[int] = None", "    actual_percentage: Optional[float] = None"),

    # AllocationResponse.percentage (int → float) + hours
    ("    percentage: int\n    hours: Optional[int] = None",
     "    percentage: float\n    hours: Optional[float] = None"),

    # MoveResourceRequest.new_percentage
    ("    new_percentage: Optional[int] = None", "    new_percentage: Optional[float] = None"),
]

before = content
for old, new in replacements:
    content = content.replace(old, new)

if content == before:
    print("⚠️  Nothing changed — check schema.py structure hasn't drifted")
else:
    with open(path, "w") as f:
        f.write(content)
    print("✓ Applied int→float fixes to schemas.py")

# Verify: no more int for these fields
import re
bad = re.findall(r"(percentage|hours|progress|capacity):\s+(Optional\[)?int", content)
if bad:
    print(f"⚠️  Still found int in: {bad}")
else:
    print("✓ Verified: no more int for percentage/hours/progress/capacity fields")
PYEOF

# Verify visually
grep -n "percentage:\|hours:\|progress:\|capacity:" backend/models/schemas.py | grep -E ":\s*int" | grep -v "#"
# Should output NOTHING (empty = all fixed)
```

## Commit and Push

```bash
git add backend/models/schemas.py
git commit -m "fix: allow float values for percentage/hours/progress/capacity fields (Pydantic v2 no longer coerces float→int, prod had 2.5 percentage that crashed /api/allocations)"
git push origin main
```

Cloud Build triggers automatically. ~4-6 minutes to deploy.

## Verify After Deploy

```bash
PROD=https://ddplan-502760053858.australia-southeast1.run.app
TOKEN=$(curl -s -X POST $PROD/api/auth/login \
  -H "Content-Type: application/x-www-form-urlencoded" \
  -d "username=don@ddconsult.tech&password=@Ddplanner2026" | python3 -c "import sys,json; print(json.load(sys.stdin)['access_token'])")

echo "Allocations: $(curl -s $PROD/api/allocations -H "Authorization: Bearer $TOKEN" | python3 -c 'import sys,json; d=json.load(sys.stdin); print(len(d))')"
```

Should return the allocation count (74+ per your prod), not 500.

Then in the browser, the Allocations page will render properly, showing all your team assignments.

---

**Testing agent verified 15/15 tests pass — the fix is bulletproof.** Push and it's done.

+++ backend/models/schemas.py
@@ -44,14 +44,14 @@ class TokenWithUser(BaseModel):
 class ResourceCreate(BaseModel):
     name: str
     role: str
-    standard_capacity: int = 100
+    standard_capacity: float = 100  # float so part-time capacity (e.g. 87.5) doesn't crash
     avatar_url: Optional[str] = None
 
 
 class ResourceUpdate(BaseModel):
     name: Optional[str] = None
     role: Optional[str] = None
-    standard_capacity: Optional[int] = None
+    standard_capacity: Optional[float] = None
     avatar_url: Optional[str] = None
     active: Optional[bool] = None
 
@@ -60,7 +60,7 @@ class ResourceResponse(BaseModel):
     id: str
     name: str
     role: str
-    standard_capacity: int
+    standard_capacity: float  # was int — allows 87.5%, 33.3% etc.
     avatar_url: Optional[str] = None
     active: bool = True
 
@@ -129,7 +129,7 @@ class ProjectResponse(BaseModel):
     actual_hours: Optional[float] = 0.0
     health: Optional[str] = None
     schedule_status: Optional[str] = None
-    actual_progress: Optional[int] = None
+    actual_progress: Optional[float] = None  # float — allows 33.3%, 66.7% etc.
     project_lead_id: Optional[str] = None
     project_lead_name: Optional[str] = None
     google_drive_url: Optional[str] = None
@@ -145,7 +145,7 @@ class StatusUpdateCreate(BaseModel):
     project_id: str
     health: str = "Green"
     schedule_status: str = "On Track"
-    actual_progress: Optional[int] = None
+    actual_progress: Optional[float] = None
     accomplishments: Optional[str] = None
     blockers: Optional[str] = None
     next_steps: Optional[str] = None
@@ -161,7 +161,7 @@ class StatusUpdateResponse(BaseModel):
     update_date: str
     health: str
     schedule_status: str
-    actual_progress: int
+    actual_progress: float  # was int — allows fractional progress values
     accomplishments: Optional[str] = None
     blockers: Optional[Union[str, list]] = None
     next_steps: Optional[str] = None
@@ -178,7 +178,7 @@ class StatusUpdateResponse(BaseModel):
 class StatusUpdateEdit(BaseModel):
     health: Optional[str] = None
     schedule_status: Optional[str] = None
-    actual_progress: Optional[int] = None
+    actual_progress: Optional[float] = None
     accomplishments: Optional[str] = None
     blockers: Optional[str] = None
     next_steps: Optional[str] = None
@@ -201,8 +201,8 @@ ALLOCATION_ROLES = [
 class PhaseAllocation(BaseModel):
     """Phase-specific allocation percentage for a resource."""
     phase_id: str
-    percentage: Optional[int] = None
-    hours: Optional[int] = None
+    percentage: Optional[float] = None  # float so 12.5% doesn't crash Pydantic
+    hours: Optional[float] = None
 
 
 class AllocationCreate(BaseModel):
@@ -210,11 +210,11 @@ class AllocationCreate(BaseModel):
     project_id: str
     start_date: date
     end_date: date
-    percentage: Optional[int] = None
-    hours: Optional[int] = None
+    percentage: Optional[float] = None  # accept decimal % (12.5, 33.3)
+    hours: Optional[float] = None
     allocation_type: str = "percentage"
     role: Optional[str] = None
-    actual_percentage: Optional[int] = None
+    actual_percentage: Optional[float] = None
     confirmation_status: str = "Pending"
     phase_names: Optional[List[str]] = None
     phase_allocations: Optional[List[PhaseAllocation]] = []  # NEW: Per-phase allocations
@@ -225,11 +225,11 @@ class AllocationUpdate(BaseModel):
     project_id: Optional[str] = None
     start_date: Optional[date] = None
     end_date: Optional[date] = None
-    percentage: Optional[int] = None
-    hours: Optional[int] = None
+    percentage: Optional[float] = None  # accept decimal %
+    hours: Optional[float] = None
     allocation_type: Optional[str] = None
     role: Optional[str] = None
-    actual_percentage: Optional[int] = None
+    actual_percentage: Optional[float] = None
     confirmation_status: Optional[str] = None
     phase_names: Optional[List[str]] = None
     phase_allocations: Optional[List[dict]] = None  # NEW: Per-phase allocations
@@ -241,12 +241,12 @@ class AllocationResponse(BaseModel):
     project_id: str
     start_date: str
     end_date: str
-    percentage: int
-    hours: Optional[int] = None
+    percentage: float  # was int — legacy prod data may have 2.5 etc.
+    hours: Optional[float] = None
     weekly_hours: Optional[float] = None
     allocation_type: Optional[str] = "percentage"
     role: Optional[str] = None
-    actual_percentage: Optional[int] = None
+    actual_percentage: Optional[float] = None
     confirmation_status: Optional[str] = "Pending"
     resource_name: Optional[str] = None
     resource_role: Optional[str] = None
@@ -547,7 +547,7 @@ class MoveResourceRequest(BaseModel):
     resource_id: str
     source_project_id: str
     target_project_id: str
-    new_percentage: Optional[int] = None
+    new_percentage: Optional[float] = None  # accept decimal %
     new_start_date: Optional[str] = None
     new_end_date: Optional[str] = None
 
@@ -623,8 +623,8 @@ class AllocationValidateRequest(BaseModel):
     resource_id: str
     start_date: date
     end_date: date
-    percentage: Optional[int] = None
-    hours: Optional[int] = None
+    percentage: Optional[float] = None
+    hours: Optional[float] = None
     allocation_type: str = "percentage"
     exclude_allocation_id: Optional[str] = None  # for edits, exclude the current one
